@@ -1,79 +1,78 @@
-﻿namespace irods_Csharp
+﻿namespace irods_Csharp;
+
+public class CollectionManager
 {
-    public class CollectionManager
+    public IrodsSession Session;
+    private readonly Path _home;
+
+    /// <summary>
+    /// Collection manager constructor.
+    /// </summary>
+    /// <param name="session">Session which contains account and connection</param>
+    /// <param name="home">Path to home directory</param>
+    public CollectionManager(IrodsSession session, string home)
     {
-        public IrodsSession Session;
-        private readonly Path _home;
+        Session = session;
+        _home = new Path(home);
+    }
 
-        /// <summary>
-        /// Collection manager constructor.
-        /// </summary>
-        /// <param name="session">Session which contains account and connection</param>
-        /// <param name="home">Path to home directory</param>
-        public CollectionManager(IrodsSession session, string home)
+    /// <summary>
+    /// Renames collection.
+    /// </summary>
+    /// <param name="source">Original name of collection</param>
+    /// <param name="target">New name of collection</param>
+    public void Rename(string source, string target)
+    {
+        Session.Rename(source, target, true);
+    }
+
+    /// <summary>
+    /// Looks for collection on server.
+    /// </summary>
+    /// <param name="path">Path to collection parent</param>
+    /// <param name="id">Id of collection, will be queried if not supplied</param>
+    /// <returns>Collection object</returns>
+    public Collection Open(string path, int id = -1)
+    {
+        if (id == -1)
         {
-            Session = session;
-            _home = new Path(home);
+            id = Session.Queries.QueryCollection(Path.First(path), Path.Last(path), true)[0].Id;
         }
+        return new Collection(new Path(path), id, this);
+    }
 
-        /// <summary>
-        /// Renames collection.
-        /// </summary>
-        /// <param name="source">Original name of collection</param>
-        /// <param name="target">New name of collection</param>
-        public void Rename(string source, string target)
+    /// <summary>
+    /// Creates collection.
+    /// </summary>
+    /// <param name="path">Path where collection should be created, including name</param>
+    public void Create(string path)
+    {
+        KeyValPair_PI mkdirRequestMsgPair = new (0, new string[0], new string[0]);
+        Packet<CollInpNew_PI> mkdirRequest = new (ApiNumberData.COLL_CREATE_AN)
         {
-            Session.Rename(source, target, true);
-        }
+            MsgBody = new CollInpNew_PI(_home + path, 0, 0, mkdirRequestMsgPair)
+        };
+        Session.SendPacket(mkdirRequest);
 
-        /// <summary>
-        /// Looks for collection on server.
-        /// </summary>
-        /// <param name="path">Path to collection parent</param>
-        /// <param name="id">Id of collection, will be queried if not supplied</param>
-        /// <returns>Collection object</returns>
-        public Collection Open(string path, int id = -1)
+        Session.ReceivePacket<None>();
+    }
+
+    /// <summary>
+    /// Removes collection.
+    /// </summary>
+    /// <param name="path">Path to collection parent</param>
+    /// <param name="recursive">Delete items of collection if they exist</param>
+    public void Remove(string path, bool recursive = true)
+    {
+        Packet<CollInpNew_PI> rmdirRequest = new (ApiNumberData.RM_COLL_AN)
         {
-            if (id == -1)
+            MsgBody = new CollInpNew_PI(_home + path, 0, 0)
             {
-                id = Session.Queries.QueryCollection(Path.First(path), Path.Last(path), true)[0].Id;
+                KeyValPair_PI = recursive ? new KeyValPair_PI(1, new[] { "recursiveOpr" }, new[] { "" }) : new KeyValPair_PI(0, new string[0], new string[0])
             }
-            return new Collection(new Path(path), id, this);
-        }
+        };
+        Session.SendPacket(rmdirRequest);
 
-        /// <summary>
-        /// Creates collection.
-        /// </summary>
-        /// <param name="path">Path where collection should be created, including name</param>
-        public void Create(string path)
-        {
-            KeyValPair_PI mkdirRequestMsgPair = new KeyValPair_PI(0, new string[0], new string[0]);
-            Packet<CollInpNew_PI> mkdirRequest = new Packet<CollInpNew_PI>(ApiNumberData.COLL_CREATE_AN)
-            {
-                MsgBody = new CollInpNew_PI(_home + path, 0, 0, mkdirRequestMsgPair)
-            };
-            Session.SendPacket(mkdirRequest);
-
-            Session.ReceivePacket<None>();
-        }
-
-        /// <summary>
-        /// Removes collection.
-        /// </summary>
-        /// <param name="path">Path to collection parent</param>
-        /// <param name="recursive">Delete items of collection if they exist</param>
-        public void Remove(string path, bool recursive = true)
-        {
-            Packet<CollInpNew_PI> rmdirRequest = new Packet<CollInpNew_PI>(ApiNumberData.RM_COLL_AN)
-            {
-                MsgBody = new CollInpNew_PI(_home + path, 0, 0)
-                {
-                    KeyValPair_PI = recursive ? new KeyValPair_PI(1, new[] { "recursiveOpr" }, new[] { "" }) : new KeyValPair_PI(0, new string[0], new string[0])
-                }
-            };
-            Session.SendPacket(rmdirRequest);
-
-            Session.ReceivePacket<CollOprStat_PI>();
-        }
+        Session.ReceivePacket<CollOprStat_PI>();
     }
 }
